requirements.txt
google-generativeai
typing-extensions
tk
ttk
pillow
